requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
shap
aif360
imbalanced-learn
groq
python-dotenv
pydantic
faker
lime
onnx
skl2onnx
scipy
matplotlib
requests
PyJWT==2.8.0
firebase-admin==6.5.0
reportlab==4.2.0
